Map defs instead of instances to firm entities.
In Polyglot, Defs are associated with the declaration AST node, e.g. each
MethodDecl node owns an instance of MethodDef.  This MethodDef is therefore
unique because there exists only one MethodDecl node for a specific method.
Instances on the other hand are associated with AST nodes that express "uses",
e.g. each Call AST node owns a MethodInstance object.  These MethodInstance
objects are not unique, i.e. different Call AST nodes reference different
MethodInstance objects.  Furthermore, each MethodInstance object contains a
reference to the appropriate MethodDef.  For some reason, Polyglot chose to
make this use-definition relationship explicit also for types.
For the mentioned reasons, it is illegal to map MethodInstances to firm entities
(because MethodInstances are not unique) and MethodDefs should be used instead. The MethodDecl would be even better but a Call node does not reference the Decl
node of the called method, so the only viable option is using the MethodDef.