change struct constructor calling convention
struct constructors now return the struct as a value. (Instead of the
old system where a struct reference was passed as argument).
This leaves the choice on how to implement struct returning to the
backend. Most backends pass in a reference anyway so the generated code
should be the same, but for small structs the compiler can now choose to
use registers.