FirmGenerator: endless loop adjustments
X10 goes on to prove that some loops run endless and then does not
require a "return" statement behind the loop. So we are forced to also
detect these loops as endless and then set current block to Bad in this
case so we not break on a missing return statement.