Allocate obstack chunks in TLM when using OctoPOS.
Push DMAs on the hardware are only allowed if the source address is a
tile-local TLM address.  Hence, as we serialize into an obstack and we want
to move the data using a DMA transfer, having the obstack chunks in TLM saves
a copy.  Also, putting runtime data structures into TLM is probably a good
idea from a performance perspective.