Revert "Reenable Monitor for static initialization synchronization."
This reverts commit a70d9307aba8ead86c29382d5e18890642a8d5ec.

This actually contains a race condition, because the code to initialize
staticMonitor itself uses staticMonitor.  Consider 3 i-lets:

1) Grabs the GC lock and allocates some stuff (takes some time).
2) Uses staticMonitor first, hence sets its state from UNINITIALIZED to
INITIALIZING.  A Monitor has a stack of Signals, so it wants to allocate
some memory.  However, i-let 1 has the GC lock, so it must wait.
3) Uses staticMonitor next and notices that its state is INITIALIZING,
so it tries to grab the InitDispatcher lock.  However, this lock *is the
static monitor* (Monitor extends Lock).  So this causes an endless
recursion and a stack overflow.