Statically allocate initStatus for static fields.
In commit a71a7c806d62ed80f437e25a1f2a518b7849a214, default allocation
for AtomicIntegers was changed to TLM memory.  This is reasonable,
because CAS does not work on global memory on the hardware.  However,
the initStatus fields for static fields must be allocated *statically*,
otherwise they will be freed on a retreat (which always happens after
the initialization of statics in the invasive mode of our runtime).  So
default allocation for AtomicIntegers is still the TLM, however, the
static AtomicIntegers generated by the static initialization code will
now be statically initialized.