Fix initialization for fields with non-trivial initializers in NativeRep classes.
The static initializer lowerer code contained a special case for fields like
Float.MAX_VALUE, which is defined as follows:

@Native("c++", "x10aux::float_utils::fromIntBits(0x7f7fffff)")
public static MAX_VALUE: Float = Float.fromIntBits(0x7f7fffff);

In this case, the static initializer lowerer assumes that static fields
with non-trivial initializing expressions, which are part of a NativeRep
class, always have a @Native annotation.  Hence, at each use, the
variable name "MAX_VALUE" is replaced with the (C++/Java) expression in
the Native annotation.  Therefore, the initializing expression is
rewritten to 0.

In the case of x10firm, we do not have @Native annotations, so we
initialized MAX_VALUE to the rewritten value (0).

This patch changes the logic so that also all accesses to these static
fields are replaced with a call to the respective get method.  In the
case of constant initializers, this results in efficient code after
inlining.