Fix race condition
The ilet_dma_remote_finish function sets the place_id, but is
not synchronized with the signal (see ilet_dma_local_finish).
This meant after a reinvade we could read from place_id, which
was not set correctly yet.

Multigrid rarely suffered from this. I extracted the PlaceCheck2 test case,
which needs 500 iterations to crash pretty reliably.

The fix was to set the place_id earlier. Due to the limitation that
ilets can only transfer two words, we now mix place_id and number
of places in one word. Somewhat ugly, but 65536 places should be enough
for InvasIC and we save sending one ilet per tile per invasion.

The ilet_dma_remote_finish is not necessary anymore, so I removed it.