Use OctoPOS' GC functionality
Use OctoPOS' garbage collector when the target platform is OctoPOS. This
enables fine-grained control over x10 memory-usage (maybe even controlled by
the agent system) and better maintainability in case of wild changes in OctoPOS
or the hardware memory-architecture. From the point of view of OctoPOS the GC
is still located in userspace and provided as a library. This library is called
octomem and, conveniently, provides exactly the same functions as defined in
x10.firm_runtime/src-c/xmalloc.h with an "octo_" prefix.